requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
greenlet==3.2.4
idna==3.11
mysqlclient==2.2.7
playwright==1.55.0
pyee==13.0.0
requests==2.32.5
RouterOS-api==0.21.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
